+++ dataforge-context/src/commonMain/kotlin/space/kscience/dataforge/context/Global.kt
@@ -1,7 +1,7 @@
 package space.kscience.dataforge.context
 
 import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.Job
 import space.kscience.dataforge.meta.Meta
 import space.kscience.dataforge.names.asName
 import kotlin.coroutines.CoroutineContext
@@ -13,43 +13,11 @@ internal expect val globalLoggerFactory: PluginFactory<out LogManager>
  * A global root context. Closing [Global] terminates the framework.
  */
 @ThreadLocal
-public object Global : Context("GLOBAL".asName(), null, Meta.EMPTY) {
-
-    override val coroutineContext: CoroutineContext = GlobalScope.coroutineContext + SupervisorJob()
-
-    /**
-     * The default logging manager
-     */
-    public val logger: LogManager by lazy { globalLoggerFactory.invoke(context = this).apply { attach(this@Global) } }
-
-    /**
-     * Closing all contexts
-     *
-     * @throws Exception
-     */
-    override fun close() {
-        logger.info { "Shutting down GLOBAL" }
-        for (ctx in contextRegistry.values) {
-            ctx.close()
-        }
-        super.close()
-    }
-
-    private val contextRegistry = HashMap<String, Context>()
-
-    /**
-     * Get previously built context
-     *
-     * @param name
-     * @return
-     */
-    public fun getContext(name: String): Context? {
-        return contextRegistry[name]
-    }
-
-    public fun context(name: String, parent: Context = this, block: ContextBuilder.() -> Unit = {}): Context =
-        ContextBuilder(parent, name).apply(block).build().also {
-            contextRegistry[name] = it
-        }
-
+private object GlobalContext : Context("GLOBAL".asName(), null, Meta.EMPTY) {
+    override val coroutineContext: CoroutineContext = GlobalScope.coroutineContext + Job()
 }
+
+public val Global: Context get() = GlobalContext
+
+public fun Context(name: String? = null, block: ContextBuilder.() -> Unit = {}): Context =
+    Global.buildContext(name, block)