+++ dataforge-io/src/commonMain/kotlin/space/kscience/dataforge/io/TaglessEnvelopeFormat.kt
@@ -197,7 +197,7 @@ public class TaglessEnvelopeFormat(
             return TaglessEnvelopeFormat(context.io, meta)
         }
 
-        private val default by lazy { invoke() }
+        private val default by lazy { invoke(context = ioContext) }
 
         override fun readPartial(input: Input): PartialEnvelope =
             default.run { readPartial(input) }