+++ dataforge-meta/src/commonMain/kotlin/space/kscience/dataforge/meta/descriptors/descriptorExtensions.kt
@@ -4,7 +4,7 @@ import space.kscience.dataforge.names.Name
 import space.kscience.dataforge.values.ValueType
 import space.kscience.dataforge.values.asValue
 
-public inline fun <reified E : Enum<E>> NodeDescriptor.enum(
+public inline fun <reified E : Enum<E>> NodeDescriptorBuilder.enum(
     key: Name,
     default: E?,
     crossinline modifier: ValueDescriptor.() -> Unit = {},