+++ dataforge-meta/src/commonTest/kotlin/space/kscience/dataforge/meta/SchemeTest.kt
@@ -1,8 +1,10 @@
 package space.kscience.dataforge.meta
 
+import space.kscience.dataforge.misc.DFExperimental
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
+@DFExperimental
 class SchemeTest {
     @Test
     fun testSchemeWrappingBeforeEdit(){
@@ -19,5 +21,16 @@ class SchemeTest {
         val config = Config()
         scheme.retarget(config)
         assertEquals(29, scheme.a)
+    }
+
+    @Test
+    fun testSchemeSubscription(){
+        val scheme = TestScheme.empty()
+        var flag: Int? = null
+        scheme.useProperty(TestScheme::a){a->
+            flag = a
+        }
+        scheme.a = 2
+        assertEquals(2, flag)
     }
 }