+++ dataforge-scripting/src/jvmTest/kotlin/space/kscience/dataforge/scripting/BuildersKtTest.kt
@@ -4,7 +4,6 @@ import space.kscience.dataforge.context.Global
 import space.kscience.dataforge.meta.get
 import space.kscience.dataforge.meta.int
 import space.kscience.dataforge.workspace.WorkspaceBuilder
-
 import space.kscience.dataforge.workspace.target
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -12,13 +11,13 @@ import kotlin.test.assertEquals
 
 class BuildersKtTest {
     @Test
-    fun checkBuilder(){
+    fun checkBuilder() {
         val workspace = WorkspaceBuilder(Global).apply {
             println("I am working")
 
-            context("test")
+            context { name("test") }
 
-            target("testTarget"){
+            target("testTarget") {
                 "a" put 12
             }
         }
@@ -29,7 +28,7 @@ class BuildersKtTest {
         val script = """
             println("I am working")
 
-            context("test")
+            context{ name("test") }
 
             target("testTarget"){
                 "a" put 12